bebob-runtime: apogee: use protocol implementation of stream parameters
