+++ libs/bebob/runtime/src/apogee.rs
@@ -2,5 +2,3 @@
 // Copyright (c) 2020 Takashi Sakamoto
 
 pub mod ensemble_model;
-
-mod apogee_ctls;
